# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Monid Finance API",
    "version": "1.0.0",
    "summary": "A Financial Datasets-compatible finance REST API backed by Monid providers.",
...
    "contact": {
      "url": "https://monid-us-finance-mcp-2026.vercel.app"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://monid-finance-api.fly.dev",
      "description": "Production (Fly.io)."
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/filings/types": {
      "get": {
        "operationId": "list_filing_types",
        "tags": [
          "Filings"
        ],
        "summary": "Available filing types",
        "description": "Returns the static filing_type enum get_filings/get_filing_items validate against (go/service/coverage.go's filingTypeOrder, Financial Datasets' own documented enum order: 10-K, 10-Q, 8-K, 20-F, 6-K). This is a fixed, hardcoded catalog derived from the same enum this server's own validation uses, not a live query, so it never calls Monid and is served at zero cost.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The five filing types this server validates filing_type against.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingTypesResponse"
                },
                "examples": {
                  "example": {
                    "value": {
                      "resource": "filing_types",
                      "filing_types": [
                        "10-K",
                        "10-Q",
                        "8-K",
                        "20-F",
                        "6-K"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/filings/items/types": {
      "get": {
        "operationId": "list_filing_item_types",
        "tags": [
          "Filings"
        ],
        "summary": "Available filing item types",
        "description": "Returns the static SEC form-instruction item catalog get_filing_items reads `item` values from (go/providers/filingitems.go's Catalogs; CatalogScope: \"Static SEC form-instruction catalog; no Monid upstream call or claim.\"). Covers only the three forms get_filing_items extracts from (10-K, 10-Q, 8-K) - 20-F and 6-K are valid get_filings filing_type values but have no item catalog here. Omitting filing_type returns all three keys; passing one narrows to that key. This is a fixed, hardcoded catalog, not a live query, so it never calls Monid and is served at zero cost.",
        "parameters": [
          {
            "name": "filing_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "10-K",
                "10-Q",
                "8-K"
              ]
            },
            "description": "Narrow the catalog to one filing type. Omit to get all three. Case-insensitive; an unrecognized value answers 400 bad_request."
          }
        ],
        "responses": {
          "200": {
            "description": "Illustrative example, truncated for readability and not a captured API sample: each form's real catalog carries every item that form's SEC instructions define.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingItemTypesResponse"
                },
                "examples": {
                  "example": {
                    "value": {
                      "10-K": [
                        {
                          "name": "Item-1",
                          "title": "Business",
                          "description": "Business"
                        },
                        {
                          "name": "Item-1A",
                          "title": "Risk Factors",
                          "description": "Risk Factors"
                        },
                        {
                          "name": "Item-7",
                          "title": "Management's Discussion and Analysis of Financial Condition and Results of Operations",
                          "description": "Management's Discussion and Analysis of Financial Condition and Results of Operations"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/financials": {
      "get": {
        "operationId": "get_all_financials",
        "tags": [
          "Financial statements"
        ],
        "summary": "Get all financial statements",
        "description": "Fetches a company's income statement, balance sheet, and cash flow statement together in one call, composed into {\"financials\": {...}} (go/service/allfinancials.go). Shares get_income_statement/get_balance_sheet/get_cash_flow_statement's own period/limit/report_period* validation (parseStatementArgs) and the same annual-vs-ttm REST default note below, but fetches the shared /equities/v1/statements + /equities/v1/filings payload once and fans the three per-kind record builds out over goroutines, rather than three separate paid calls: composing this from the three individual routes would cost 3x.\n\nperiod default note: like the three individual statement routes, this REST route always fills a period value before calling the underlying capability, so an omitted `period` query parameter defaults to \"annual\" here, not the capability's own \"ttm\" default. Pass `period=ttm` explicitly for trailing-twelve-month rows.\n\ncik and filing_date* filtering are not forwarded by this REST route (see docs/openapi-notes.md's documented REST-vs-tool deviations for the three individual statement routes; the same gap applies here).",
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Uppercase-normalized ticker symbol (1-20 letters/digits/dots/hyphens)."
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "annual",
                "quarterly",
                "ttm"
              ],
              "default": "annual"
            },
            "description": "Reporting period. REST-level default is \"annual\" (see description); the capability's own default is \"ttm\"."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 4
            },
            "description": "Report periods per statement kind, most recent first (default 4, max 100)."
          },
          {
            "name": "as_reported",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Accepted for FD parameter parity but always rejected: as_reported=true answers 400 bad_request (\"as_reported is not supported; use the normalized financials endpoint\") at zero cost."
          },
          {
            "name": "report_period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Exact report_period date filter (YYYY-MM-DD)."
          },
          {
            "name": "report_period_gte",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "report_period >= this date (YYYY-MM-DD, inclusive)."
          },
          {
            "name": "report_period_lte",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "report_period <= this date (YYYY-MM-DD, inclusive)."
          },
          {
            "name": "report_period_gt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "report_period > this date (YYYY-MM-DD, exclusive)."
          },
          {
            "name": "report_period_lt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "report_period < this date (YYYY-MM-DD, exclusive)."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Not used: this route is never paginated (see AllFinancialsResponse)."
          }
        ],
        "responses": {
          "200": {
            "description": "Composed income/balance/cash-flow statements for one ticker. Real AAPL FY2025 annual values captured from our own live API (docs/our-live-samples/income_statements.json, balance_sheets.json, cash_flow.json).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllFinancialsResponse"
                },
                "examples": {
                  "example": {
                    "value": {
                      "financials": {
                        "income_statements": [
                          {
                            "ticker": "AAPL",
                            "report_period": "2025-09-30",
                            "fiscal_period": "FY2025",
                            "period": "annual",
                            "revenue": 416161000000,
                            "cost_of_revenue": 220960000000,
                            "gross_profit": 195201000000,
                            "operating_expense": 62151000000,
                            "research_and_development": 34550000000,
                            "operating_income": 133050000000,
                            "ebit": 133050000000,
                            "income_tax_expense": 20719000000,
                            "net_income": 112010000000,
                            "net_income_common_stock": 112010000000,
                            "earnings_per_share": 7.49,
                            "earnings_per_share_diluted": 7.46,
                            "weighted_average_shares": 14948500000,
                            "weighted_average_shares_diluted": 14948500000
                          }
                        ],
                        "balance_sheets": [
                          {
                            "ticker": "AAPL",
                            "report_period": "2025-09-30",
                            "fiscal_period": "FY2025",
                            "period": "annual",
                            "total_assets": 359241000000,
                            "current_assets": 147957000000,
                            "cash_and_equivalents": 35934000000,
                            "inventory": 5718000000,
                            "trade_and_non_trade_receivables": 39777000000,
                            "non_current_assets": 211284000000,
                            "total_liabilities": 285508000000,
                            "current_liabilities": 165631000000,
                            "current_debt": 20329000000,
                            "non_current_liabilities": 119877000000,
                            "non_current_debt": 78328000000,
                            "shareholders_equity": 73733000000,
                            "retained_earnings": -14264000000
                          }
                        ],
                        "cash_flow_statements": [
                          {
                            "ticker": "AAPL",
                            "report_period": "2025-09-30",
                            "fiscal_period": "FY2025",
                            "period": "annual",
                            "net_cash_flow_from_operations": 111482000000,
                            "net_cash_flow_from_investing": 27910000000,
                            "dividends_and_other_cash_distributions": -15421000000,
                            "net_cash_flow_from_financing": -120686000000,
                            "change_in_cash_and_equivalents": 18706000000,
                            "ending_cash_balance": 35934000000,
                            "free_cash_flow": 98767000000
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/financials/search/line-items": {
      "post": {
        "operationId": "search_line_items",
        "tags": [
          "Financial statements"
        ],
        "summary": "Search financial statement line items across tickers",
        "description": "Searches an arbitrary set of statement field names (from IncomeStatement/BalanceSheet/CashFlowStatement) across an arbitrary, capped set of tickers, merging one row per (ticker, report_period) (go/service/searchlineitems.go). Reuses the same statement-matrix parsing get_income_statement/get_balance_sheet/get_cash_flow_statement build on, fetching all three statement kinds for every requested ticker and returning only the caller's requested field names. `limit` bounds how many report periods each ticker contributes (most recent first) - Financial Datasets' own schema does not say whether limit is per-ticker or global, and per-ticker is what every other limit in this API means. `tickers` is capped at 5 entries: each ticker is its own paid /equities/v1/statements call, so without a cap one request could turn into an unbounded spend.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchLineItemsRequest"
              },
              "examples": {
                "example": {
                  "value": {
                    "line_items": [
                      "revenue",
                      "free_cash_flow"
                    ],
                    "tickers": [
                      "AAPL"
                    ],
                    "period": "annual",
                    "limit": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "One row per (ticker, report_period). Real AAPL FY2025 revenue/free_cash_flow values captured from our own live API (docs/our-live-samples/income_statements.json, cash_flow.json). Never paginated (no next_page_url; use limit to bound the result set per ticker).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemsSearchResponse"
                },
                "examples": {
                  "example": {
                    "value": {
                      "search_results": [
                        {
                          "ticker": "AAPL",
                          "report_period": "2025-09-30",
                          "period": "annual",
                          "revenue": 416161000000,
                          "free_cash_flow": 98767000000
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "The caller's own Monid API key (get one at https://monid.ai). Passed straight through to Monid on every call, so usage bills the caller's own wallet — this is never a shared server-side credential, and the server never logs or stores it. Missing, empty, or malformed values answer 401 unauthorized before any paid call is attempted."
      }
    },
    "schemas": {
...
      "FilingTypesResponse": {
        "type": "object",
        "title": "FilingTypesResponse",
        "description": "Static enum response for /filings/types: the filing_type values get_filings/get_filing_items validate against (go/service/coverage.go's filingTypeOrder, Financial Datasets' own documented enum order). Makes no Monid call.",
        "properties": {
          "resource": {
            "type": "string",
            "description": "Label identifying what this catalog is a list of."
          },
          "filing_types": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "10-K",
                "10-Q",
                "8-K",
                "20-F",
                "6-K"
              ]
            },
            "description": "Accepted filing_type values, in Financial Datasets' own documented enum order."
          }
        },
        "required": [
          "resource",
          "filing_types"
        ]
      },
...
      "FilingItemTypesResponse": {
        "type": "object",
        "title": "FilingItemTypesResponse",
        "description": "Static SEC form-instruction item catalog for /filings/items/types (go/providers/filingitems.go's Catalogs; CatalogScope: \"Static SEC form-instruction catalog; no Monid upstream call or claim.\"). Keyed by filing type (\"10-K\", \"10-Q\", \"8-K\" - the only three forms get_filing_items extracts from; 20-F/6-K are not covered here even though they are valid get_filings filing_type values). Omitting `filing_type` returns all three keys; passing one returns only that key. Makes no Monid call.",
        "additionalProperties": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/FilingItemTypeEntry"
          }
        }
      },
      "AllFinancialsResponse": {
        "type": "object",
        "title": "AllFinancialsResponse",
        "description": "get_all_financials envelope: {\"financials\": {\"income_statements\": [...], \"balance_sheets\": [...], \"cash_flow_statements\": [...]}}. Never paginated (no next_page_url); use `limit` to bound each statement kind's record count. Composes the same three fetches get_income_statement/get_balance_sheet/get_cash_flow_statement each make individually, but fetches the shared /equities/v1/statements + /equities/v1/filings payload once and fans the three record builds out over goroutines, rather than three separate paid calls (go/service/allfinancials.go).",
        "properties": {
          "financials": {
            "type": "object",
            "properties": {
              "income_statements": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/IncomeStatement"
                }
              },
              "balance_sheets": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BalanceSheet"
                }
              },
              "cash_flow_statements": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CashFlowStatement"
                }
              }
            },
            "required": [
              "income_statements",
              "balance_sheets",
              "cash_flow_statements"
            ]
          }
        },
        "required": [
          "financials"
        ]
      },
...
      "LineItemsSearchResponse": {
        "type": "object",
        "title": "LineItemsSearchResponse",
        "description": "search_line_items envelope: {\"search_results\": [...]}. Never paginated (no next_page_url; use limit to bound the result set per ticker).",
        "properties": {
          "search_results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineItemSearchRow"
            }
          }
        },
        "required": [
          "search_results"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "400 Bad Request. Validation failed before any paid Monid call was made (bad_request), or the request is well-formed but names a capability this server deliberately rejects (unsupported), e.g. as_reported=true or a non-USD currency.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "bad_request": {
                "summary": "Validation failure",
                "value": {
                  "error": "bad_request",
                  "message": "ticker must be 1-20 letters, digits, dots, or hyphens"
                }
              },
              "unsupported": {
                "summary": "Deliberately rejected capability",
                "value": {
                  "error": "unsupported",
                  "message": "as_reported is not supported by the Monid-backed server; as_reported=True cannot be answered honestly."
                }
              },
              "invalid_cursor": {
                "summary": "Malformed pagination cursor",
                "value": {
                  "error": "invalid_cursor",
                  "message": "cursor is not a valid opaque pagination token"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "401 Unauthorized. The X-API-KEY header was missing, empty, or contained control characters, or (if the operator configured an allowlist) the key is not on it. This check runs before any Monid call.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "error": "unauthorized",
                  "message": "Missing or invalid API key."
                }
              }
            }
          }
        }
      },
      "PaymentRequired": {
        "description": "402 Payment Required. The caller's own Monid wallet could not cover the call (Monid returned a blocked/insufficient-funds result).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "payment_required": {
                "value": {
                  "error": "payment_required",
                  "message": "Monid run blocked: insufficient wallet balance."
                }
              }
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "429 Too Many Requests. Per-API-key token-bucket rate limit exceeded (default 60/min per key; 10/min for keyless demo traffic, where enabled).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "rate_limited": {
                "value": {
                  "error": "rate_limited",
                  "message": "Rate limit exceeded. Retry shortly."
                }
              }
            }
          }
        }
      },
      "BadGateway": {
        "description": "502 Bad Gateway. The upstream Monid provider call failed for a reason other than auth/funds/timeout (upstream_error), or its payload no longer matches the shape this server parses (upstream_schema_changed).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "upstream_error": {
                "summary": "Generic upstream failure",
                "value": {
                  "error": "upstream_error",
                  "message": "provider returned HTTP 500"
                }
              },
              "upstream_schema_changed": {
                "summary": "Upstream payload shape drifted",
                "value": {
                  "error": "upstream_schema_changed",
                  "message": "provider payload is not valid JSON"
                }
              }
            }
          }
        }
      },
      "GatewayTimeout": {
        "description": "504 Gateway Timeout. The upstream Monid provider call timed out.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "upstream_timeout": {
                "value": {
                  "error": "upstream_timeout",
                  "message": "provider request timed out"
                }
              }
            }
          }
        }
      }
    }
  }
}